refactor: (data-response) add data type with template
