+++ src/main/java/connectingstar/tars/habit/response/HabitHistoryDateGetResponseV1.java
@@ -7,10 +7,13 @@ import java.time.LocalDateTime;
 
 /**
  * GET /habit/history/date API의 response
+ *
+ * @Deprecated use HabitDailyTrackingResponse instead.
  */
+@Deprecated
 @Getter
 @JsonInclude(JsonInclude.Include.NON_NULL)
-public class HabitHistoryDateGetResponse {
+public class HabitHistoryDateGetResponseV1 {
     /**
      * 습관 수행 날짜
      */
@@ -37,7 +40,7 @@ public class HabitHistoryDateGetResponse {
      */
     private final String review;
 
-    public HabitHistoryDateGetResponse(LocalDateTime runDate, String runPlace, Integer runValue, String behaviorUnit, String review) {
+    public HabitHistoryDateGetResponseV1(LocalDateTime runDate, String runPlace, Integer runValue, String behaviorUnit, String review) {
         this.runDate = runDate;
         this.runPlace = runPlace;
         this.runValue = runValue;
